Start developing against AssertJ 4.x (#6)
Motivation
----------

Eclipse Collections released version 13.0 which compiles against JDK 17. This allows us to baseline on AssertJ 4.x development which also targets JDK 17. This PR updates the project to support this. It also adds the JPMS module info files necessary to make the library a module.

Modifications
-------------

The largest modification is moving the tests to a dedicated package to avoid the split package problem across modules. There are now two module-info files: one for the main and one for the test. The test module is open by design so JUnit can find the tests. I updated the POM file to point to the 4.0.0-M1 version of AssertJ core. 

Result
------

Development can now proceed tracking all of the latest from AssertJ 4.x